fix: move tests to test/ directory for CI coverage pickup
CI runs pytest from test/, not tests/. Moved all folder import tests
and merged fixtures into the existing test/conftest.py.
